+++ core/src/main/kotlin/fr/xgouchet/elmyr/annotation/AdvancedForgery.kt
@@ -1,0 +1,30 @@
+package fr.xgouchet.elmyr.annotation
+
+/**
+ * Mark a field, property or method parameter as an advanced forgery.
+ *
+ * Advanced allow you to provide a list of ways the forgery can be created.
+ * Only one of each parameter can be provided.
+ *
+ * @param string describes all the way a String can be forged
+ * @param regex describes all the way a String can be forged with regex
+ * @param int describes all the way a Int can be forged
+ * @param long describes all the way a Long can be forged
+ * @param float describes all the way a Float can be forged
+ * @param double describes all the way a Double can be forged
+ */
+@Target(
+    AnnotationTarget.VALUE_PARAMETER,
+    AnnotationTarget.FIELD,
+    AnnotationTarget.PROPERTY
+)
+@Retention(AnnotationRetention.RUNTIME)
+@MustBeDocumented
+annotation class AdvancedForgery(
+    val string: Array<StringForgery> = emptyArray(),
+    val int: Array<IntForgery> = emptyArray(),
+    val long: Array<LongForgery> = emptyArray(),
+    val float: Array<FloatForgery> = emptyArray(),
+    val double: Array<DoubleForgery> = emptyArray(),
+    val map: Array<MapForgery> = emptyArray()
+)